backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]>=3.1
Pillow>=10.0
django-allauth>=0.57
requests>=2.28
gunicorn>=21.0
whitenoise>=6.6
